feat: 병원 정보 조회 기능 및 게시판 비밀글 기능 구현, ObjectMapper 설정 수정
- AI 기능 구현하면서 ObjectMapper 을 Bean 으로 직접 등록했었는데, 이때문에 SpringBoot 에서 기본적으로 주입하는 ObjectMapper 사용이 안됨 -> LocalDateTime 타입에 대한 Json 직렬화 시 문제 발생 → Bean 제거하여 해결
- 게시판 상세 조회 시 비밀글 접근 기능 추가 구현 + 게시판 서비스 및 컨트롤러 전반 리팩터링
- 병원 정보 조회 기능(반경 내 조건에 맞는 병원 요약 정보 조회 API, 병원 id 로 상세정보 조회 API) 구현

+++ src/main/java/com/ssafy/home/domain/board/dto/CommentResponseDto.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CommentResponseDto {
 	private Long commentId;				// 댓글 ID
-	private Long userId;				// 작성자
+	private int userId;				// 작성자
 	private Long parentId;				// 부모 댓글 ID (NULL이면 최상위 댓글)
 	private String content;				// 댓글 내용
 	private LocalDateTime createdAt;	// 작성일시
